Dispatcher can broadcast to a lambda
Trying to figure out what the dispatcher should do and how it should do it. It's a little tricky since in order to make the TDD steps the dispatcher needs to affect the outside world. It seems like I am making jumps that are too big but I can't just add a subscriber because that has no effect on the subscriber.