Dispatch empty event and emitted event
In the move towards an event based system I upgraded the dispatcher to take a Event objects. I'd like to put some more information on these so when two adventurers want to talk they can pass what they say via the event from this. I'm making this up as I go along so we'll see if this can be expanded upon to included messages. Eventually even other interactions.