Create the GenericDispatcher
The GenericDispatcher should be able to replace the current Dispatcher but support both Strings and Events with fewer methods. Sadly, it won't be able to support String and Events at the same time. There will need to be a different GenericDispatcher for each type in that scenario.

The next step is to have the GenericDispatcher replace the Dispatcher in a refactoring switcheroo.